refactor: watch service
